@@ -1,109 +1,74 @@
 const skillData = [
-    {
-         type: "Frontend Development",
-         aboutlanguages: "Languages:",
-         languages: [
-              "HTML5, ",
-              "CSS3, ",
-              "Sass, ",
-              "JavaScript(ES6), ",
-              "React",
-          ],
-          abouttools: "Tools I Use:",
-          tools: [
-              "GitHub, ",
-              "Gitlab, ",
-              "Bootstrap, ",
-              "React Spring, ",
-              "Particles.js, ",
-              "Figma, ",
-              "Netlify",
-         ],
-    },
-    {
-          type: "Backend Development",
-          aboutlanguages: "Languages:",
-          languages: [
-               "NodeJs, ",
-               "NoSQL, ",
-               "SQL",
-           ],
-          abouttools: "Tools I Use:",
-          tools: [
-               "Express, ",
-               "MongoDB, ",
-               "PostgreSQL, ",
-               "Heroku, ",
-               "Github",
-          ],
-     },
-     {
-          type: "Project Management",
-          aboutpractices: "Practices:",
-          practices: [
-               "Agile Methods, ",
-               "Scrum, ",
-               "Kanban, ",
-               "Design Thinking",
-          ],
-          abouttools: "Tools I Use:",
-          tools: [
-               "Trello, ",
-               "Airtable, ",
-               "Zapier, ",
-               "Monday, ",
-               "Miro, ",
-               "Slack, ",
-               "Discord, ", 
-               "MS Teams",              
-          ],          
-     },
-
-
-
-
-
-//     {
-//          type: "Frontend Development",
-//          skill: ""
-//     },
-//     {
-//          type: "Frontend Development",
-//          skill: ""
-//     },
-//     {
-//          type: "Frontend Development",
-//          skill: ""
-//     },
-   
-//     {
-//          type: "Backend Development",
-//          skill: ""
-//     },
-//     {
-//          type: "Backend Development",
-//          skill: ""
-//     },
-//     {
-//          type: "Backend Development",
-//          skill: ""
-//     },
-//     {
-//          type: "Project Management",
-//          skill: "Agile Methods, SCRUM"
-//     },
-//     {
-//          type: "Project Management",
-//          skill: "Design Thinking"
-//     },
-//     {
-//          type: "Project Management",
-//          skill: "Report Writing & Documentation"
-//     },
-//     {
-//          type: "Project Management",
-//          skill: "Planning & Organization"
-//     }
-]
+  {
+    type: "Languages",
+    items: [
+      { name: "HTML5", icon: "/assets/icons/html-icon.svg" },
+      { name: "CSS3", icon: "/assets/icons/css-icon.svg" },
+      { name: "Sass/SCSS", icon: "/assets/icons/sass-icon.svg" },
+      { name: "JavaScript", icon: "/assets/icons/javascript-icon.svg" },
+    ],
+  },
+  {
+    type: "Tools & Libraries",
+    items: [
+      { name: "React", icon: "/assets/icons/react-icon.svg" },
+      { name: "Next.js", icon: "/assets/icons/nextjs-icon.svg" },
+      { name: "Webpack", icon: "/assets/icons/webpack-icon.svg" },
+      { name: "Node.js", icon: "/assets/icons/nodejs-icon.svg" },
+      { name: "Express.js", icon: "/assets/icons/expressjs-icon.svg" },
+      { name: "SQL", icon: "/assets/icons/sql-icon.svg" },
+      { name: "MongoDB", icon: "/assets/icons/mongodb-icon.svg" },
+      { name: "Drupal", icon: "/assets/icons/drupal-icon.png" },
+      { name: "Twig", icon: "/assets/icons/twig-icon.png" },
+      { name: "Wordpress", icon: "/assets/icons/wordpress-icon.png" },
+      { name: "Bootstrap", icon: "/assets/icons/bootstrap-icon.svg" },
+      { name: "D3", icon: "/assets/icons/d3-icon.svg" },
+      { name: "Mapbox", icon: "/assets/icons/mapbox-icon.svg" },
+      { name: "React Spring", icon: "/assets/icons/reactspring-icon.png" },
+      { name: "Git", icon: "/assets/icons/git-icon.svg" },
+      { name: "GitHub", icon: "/assets/icons/github-icon.svg" },
+      { name: "GitLab", icon: "/assets/icons/gitlab-icon.svg" },
+      { name: "Docker", icon: "/assets/icons/docker-icon.svg" },
+      { name: "Jira", icon: "/assets/icons/jira-icon.svg" },
+      { name: "Figma", icon: "/assets/icons/figma-icon.svg" },
+      {
+        name: "Adobe Illustrator",
+        icon: "/assets/icons/adobeillustrator-icon.svg",
+      },
+      { name: "Zapier", icon: "/assets/icons/zapier-icon.svg" },
+      { name: "Airtable", icon: "/assets/icons/airtable-icon.svg" },
+    ],
+  },
+  {
+    type: "Project Management",
+    items: [
+      { name: "Agile Methods" },
+      { name: "Scrum" },
+      { name: "Kanban" },
+      { name: "Design Thinking" },
+      { name: "Team Work" },
+      { name: "Planning & Organization" },
+      { name: "Conducting Interviews" },
+      { name: "Time Management" },
+      { name: "Moderation & Presentation" },
+      { name: "Organizational Development" },
+    ],
+  },
+  {
+    type: "Topics",
+    items: [
+      { name: "Web Development" },
+      { name: "Software Development" },
+      { name: "Web Accessibility" },
+      { name: "Intersectional Feminism" },
+      { name: "Social & Environmental Justice" },
+      { name: "Climate Change & Migration" },
+      { name: "Political Science & International Development" },
+      { name: "Peace & Conflict Studies" },
+      { name: "Citizen Participation" },
+      { name: "Advocacy on Children´s Rights" },
+    ],
+  },
+];
 
 export default skillData;
